make spanned_bytes private
we need this for proper allocation for set schema and
set data using schema.

the value reported can be confusing if users try to draw
conclusions about the layout of a node from this value reported
by their schema. so we are keeping it private

I also changed DataType::equal and DataType::equals, and
Schema::equal to DataType::equals