enh: add ENABLE_DOCS build system option
ENABLE_DOCS (default = ON) controls if the documentation targets
are created.

The target creation is still guarded by if doxygen and sphinx are found.
If ENABLE_DOCS = ON, and doxygen and sphinx are found, you can build
conduit's documentation using the docs target (`make docs`) and
the install target (`make install`) will also build and install the
docs.