bugfix: improve string formatting for fp numbers
added a helper conduit::utils::float64_to_string()

changed conduit::DataArray::to_json() to use
this method for the float32 and float64 cases.

This strikes a balance between creating a readable
string that preserves enough info to reconstruct
data from floating numbers from (pure) json.

for example, want:

float v = 30.0;

to be printed as "30.0", not "30", so we can
consume it as floating point number not an integer.

I played with several c++ stream options, but
ultimately snprintf using %g provided 90% of what
we needed, so I used this and added a little more logic.