enh: gap methods for c types for Node::set
These set methods are used to fill out the interface for cases where
any of the native c types are not mapped 1-1 to our to bit width style
types.

Windows is one important case where this happens. Both long and int
represent 32-bit integers, and long long is used as the type for 64-bit
integers. In this case the long and int are aliased types -- we want
to support both via overloaded "set" functions, however one of the types
is already used as the underlying type for set(int32 ). When int is
selected as int32, Visual Studio needs an explicit method to disambiguate
the long case.