rework gen constructors to avoid bad overloads on windows
visual studio will match const std::string& to a void * arg, even over
an overloaded variant that takes const std::string& in the same
position.

Simplifed the generator, to avoid problems with this.
Added getters and setters to the generator class, as an
alterative to setting all options during construction