conduit/c: use a type safe alias for `conduit_{datatype,node}`
Using `typedef void` meant that the APIs were all taking `void*` which
in turn means that any pointer is valid to pass to these types. Instead,
use a unique opaque type so that callers must either do the unsafe
casting themselves or use the `cpp_to_c` APIs to do the casting the
intended way.

The actual definition of the structs themselves are empty; they are just
used it as a way to smuggle the C++ pointer around under an opaque name.

Fixes: #782