Refactor coordset/topology combine stuff
Uniform/Rectilinear/Structured meshes require knowledge about the
topology and coordset at the same time. Was able to get around this with
the first two but Structured needs access to the topology node when
combining the coordsets.

This also fixed a bug in the uniform combine tests.

Now there is no longer a coordset::combine function so the pointmerge
tests have been disabled for the time being.