Move iterate_elements into its own header
Cleaned up the code a bit to move the lambdas that were defined inline
out into their own functions under an "impl" namespace.
This header is not currently installed with Conduit as it is only being
used in the patition and now flatten implementations.
Would probably want to cleanup the interface a bit before exposing to
users - there's a lot of code in the header (templated functions).