cmake: use the commands that get the git information directly (#970)
Asking `git describe` to `--match` nothing is just a long-winded way of
asking for the hash. Instead, ask for the hash and dirty status
directly.

Note that this is not ideal anyways as CMake can have stale information
as the build is not regenerated if the working tree changes at all.