Use std::function for memory handler callbacks (#1366)
Change Conduit's memcpy, memset, alloc, and free handlers from function pointers to std::function.  This will allow these handlers to have local state, increasing the flexibility and power to allow on-the-fly creation of new handlers.  The change is backwards-compatible because you can assign a function pointer to a std::function.

---------

Co-authored-by: Cyrus Harrison <cyrush@llnl.gov>